- added getSymbol method to TextField for accessing to aligned/formatted text - core::getPackage (for android yet) - added short syntax version of actor->addEventListener(TouchEvent::CLICK, ...); - updated to latest SDL - minor

+++ oxygine/src/math/ScalarMath.h
@@ -36,7 +36,7 @@ namespace oxygine
 #ifdef IW_SYS_MATH_H
             return fabsf(s);
 #else
-            return fabs(s);
+            return fabsf(s);
 #endif
         }
